Fix task接口更新 fansq

+++ service-task/src/main/java/com/github/hollykunge/security/task/entity/LarkFile.java
@@ -3,8 +3,8 @@ package com.github.hollykunge.security.task.entity;
 import com.github.hollykunge.security.common.entity.BaseEntity;
 import lombok.Data;
 
-import java.util.Date;
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Table;
 
 /**
  * @author  fansq
@@ -28,7 +28,7 @@ public class LarkFile extends BaseEntity {
     private String extension;
 
     @Column(name = "FILE_SIZE")
-    private Short fileSize;
+    private Integer fileSize;
 
     @Column(name = "OBJECT_TYPE")
     private String objectType;
@@ -46,7 +46,7 @@ public class LarkFile extends BaseEntity {
     private String createBy;
 
     @Column(name = "DOWNLOADS")
-    private Short downloads;
+    private Integer downloads;
 
     @Column(name = "EXTRA")
     private String extra;
